refactor: organize title faker directory
